refactor: view 관련 클래스 view 패키지로 이동
