--- nextstep-mvc/src/main/java/nextstep/mvc/tobe/JsonView.java
@@ -1,12 +1,0 @@
-package nextstep.mvc.tobe;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.Map;
-
-public class JsonView implements View {
-    @Override
-    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-
-    }
-}
