+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/AnnotationHandlerMapping.java
@@ -2,7 +2,7 @@ package nextstep.mvc.tobe.handlermapping;
 
 import com.google.common.collect.Maps;
 import nextstep.mvc.HandlerMapping;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
