+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JsonView.java
@@ -1,18 +1,14 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
+
+import nextstep.mvc.tobe.View;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Map;
 
-public class RedirectView implements View {
-    private final String redirectUrl;
-
-    public RedirectView(String redirectUrl) {
-        this.redirectUrl = redirectUrl;
-    }
-
+public class JsonView implements View {
     @Override
     public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        response.sendRedirect(redirectUrl);
+
     }
 }
