+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/RedirectView.java
@@ -1,0 +1,20 @@
+package nextstep.mvc.tobe.view;
+
+import nextstep.mvc.tobe.View;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Map;
+
+public class RedirectView implements View {
+    private final String redirectUrl;
+
+    public RedirectView(String redirectUrl) {
+        this.redirectUrl = redirectUrl;
+    }
+
+    @Override
+    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
+        response.sendRedirect(redirectUrl);
+    }
+}
