+++ nextstep-mvc/src/test/java/samples/MyUserController.java
@@ -1,7 +1,7 @@
 package samples;
 
 import nextstep.db.DataBase;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
